requirements.txt
absl-py==1.4.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.3.4
astroid==3.3.9
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bash==0.6
beautifulsoup4==4.12.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.3.0
catboost==1.2.8
certifi==2021.5.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cssselect==1.2.0
cv==1.0.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distlib==0.3.2
Django==3.2.4
dnspython==2.7.0
docker==5.0.0
docker-pycreds==0.4.0
dominate==2.9.1
email_validator==2.2.0
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.0.12
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
fqdn==1.5.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.51.3
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==2.10
imageio==2.37.0
importlib-metadata==6.1.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
instaclient==2.9.20
instagram-py==2.0.7
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.6
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jax==0.4.6
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.2.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keras==3.9.2
kiwisolver==1.4.7
libclang==15.0.6.1
lightgbm==4.6.0
lolcat==1.4
lxml==4.9.2
Markdown==3.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mechanize==0.4.5
mistune==3.1.3
ml-dtypes==0.4.1
more-itertools==9.1.0
MouseInfo==0.1.3
mypy_extensions==1.1.0
mysql-connector-python==8.0.25
namex==0.0.9
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.4.1
notebook_shim==0.2.4
numpy==2.0.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.15.0
orjson==3.10.18
overrides==7.7.0
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
phonenumbers==8.12.45
pickle-mixin==1.0.2
Pillow==9.3.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.19.1
pylint==3.3.6
Pympler==1.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytweening==1.0.4
pytz==2021.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0
pyzmq==26.4.0
Quandl==3.7.0
referencing==0.36.2
regex==2023.3.23
requests==2.32.3
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.4
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.13.1
seaborn==0.13.2
selenium==3.141.0
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==2.27.0
setproctitle==1.3.6
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.4
SQLAlchemy==2.0.41
sqlparse==0.4.1
stack-data==0.6.3
starlette==0.46.2
stem==1.8.0
streamlit==1.45.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.2.0
terminado==0.18.1
threadpoolctl==3.1.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==3.4.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tor==1.0.0
torch==1.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
validators==0.35.0
virtualenv==20.4.7
visdom==0.2.4
wandb==0.19.10
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.1
WTForms==3.2.1
xgboost==2.1.4
zipp==3.15.0
